Run PackageManager init when using command line
When using the command line to import a package hasGui (false) is being passed as the loadPackagesConfig arg to PackageManager.init()

This means that TaxonomyPackage.json is never initialized. 

Then when the PackageManger.save() is called on a newly imported package it tries to write to TaxonomyPackage.json which doesn't exist, throwing an error.